Add REFERENCES keyword where possible, align columns definitions.
When you have foreign keys, it is better to use the REFERENCES keyword,
in order to enforce "exists" relationships between tables.

See: https://www.sqlite.org/foreignkeys.html

Cannot use REFERENCES for fk_id_kara in the table of the userdata database,
because the kara table is in another database.

I've aligned the columns definitions using the following function:
3 spaces after the longer column name (arbitrary choice).