Standardization of columns that are used to store a date in the databases, and remove some prefixes.
created_at: Set the date to now() at INSERT
update_at: Set the date to now() at INSERT, and update the date to now() at UPDATE
banned_at: Set the date to the banned date

Remove prefixes wl, bl. In the context of the table blacklist or whitelist, adding these
prefixes is redundant.