Added backend checks and options for time quota (#109)
New options :
- EngineTimePerUser (defaults to 240, so 4 minutes)
- EngineQuotaType (defaults to 1. 0 = none, 1 = songs, 2 = time)

The WS message has been changed to "quotaAvailableUpdated" and serves
the quotaLeft and quotaType arguments (along with the usual username
argument)

quotaLeft can be null if in admin mode or if quotatype = 0. In this case
frontends should display an infinite character.

Still need to add the new settings to the API