Fixed tests with new validators
Removed negative order test for .kara validation as users are not
supposed to be that stupid and besides it could make for a funny karaoke
with a -1 order for openings before the real opening and... ahem.

To really fix this we should need another validator which accepts
negative numbers. My fault for trying to use the same validator for both
the API and kara files. Bleh.

I HATE TESTS. :3