Reworked order in karaoke list
- A new column has been added to the all_karas materialized view. It's a
COALESCE of serie and singer. if serie is null, singer is used instead.
This new column allows the database to mix singers and series in the
ordered list
- There is no difference anymore between lower/uppercase letters and
letters with accents in the ordered karaoke lists. Which means series
beginning with an accented letter are correctly placed in the list