fix admin drag'n drop karaline for reordering them
Note : Playlist state (kara/current/public/blc/blc critterias/ ...) need to be rework as separate component and the main interface just have to use the right component base on the panel state.

This kind of rework allow us to clearly separate all those different behavior.

But for now it works ^^ so have fun