remove ReactRef handling for Karaline (back to native DOM JS)
React Ref conflict with the sortable component.
When SortDrag start some weird re-render occurs and freeze the interface

The replacement solution is based on kid injection data attribute on karaline and a simple DOM request to find element before apply the scroll. Same result but with less pain an bug ^^